requirements.txt
fastapi
uvicorn
pymongo
google-generativeai
python-dotenv
google-cloud-texttospeech